GOSSIP-15 avoid busy loop (ChiaHung Lin via egc)
